[#88] Sự khác biệt về kích thước Struct giữa C++ và C# trên x86

+++ engine/pak.h
@@ -27,7 +27,7 @@ namespace fs = std::filesystem;
 struct CompressedFileInfo {
 	int index;
 	char* id;
-	unsigned long idValue;
+	unsigned long long idValue;
 	char* time;
 	char* fileName;
 	int size;
